No mocking of RouteResult
More reliable results we can get if we create proper instance of
RouteResult. This class has two states - when matched routing is found
and when routing is failed. It is hard to properly mock internal state
of the object with all public methods. THat's why it's easier to use
static method to create the object with desired state.